refactor: retarget display builder owner
